// netflix-clone/src/types/movie.ts
export interface Movie {
  id: number
  title?: string
  name?: string
  original_title?: string
  original_name?: string
  overview: string
  poster_path: string | null
  backdrop_path: string | null
  media_type?: string
  adult: boolean
  original_language: string
  genre_ids: number[]
  popularity: number
  release_date?: string
  first_air_date?: string
  video?: boolean
  vote_average: number
  vote_count: number
}

export interface MovieResponse {
  page: number
  results: Movie[]
  total_pages: number
  total_results: number
}

export interface Genre {
  id: number
  name: string
}

export interface MovieDetails extends Movie {
  genres: Genre[]
  runtime?: number
  status: string
  tagline?: string
  budget?: number
  revenue?: number
  production_companies: {
    id: number
    name: string
    logo_path: string | null
    origin_country: string
  }[]
}